+++ src/invoices/InvoicesList/hooks/useBuildQuery/useBuildQuery.js
@@ -31,9 +31,9 @@ export const useBuildQuery = () => {
       [FILTERS.APPROVAL_DATE]: buildDateTimeRangeQuery.bind(null, [FILTERS.APPROVAL_DATE]),
       [FILTERS.PAYMENT_DATE]: buildDateTimeRangeQuery.bind(null, [FILTERS.PAYMENT_DATE]),
       [FILTERS.TAGS]: (filterValue) => {
-        const value = Array.isArray(filterValue) ? filterValue.join('" or "') : filterValue;
+        const value = Array.isArray(filterValue) ? `(${filterValue.map(v => `"*${v}*"`).join(' or ')})` : `"*${filterValue}*"`;
 
-        return `(${FILTERS.TAGS}=("${value}") or invoiceLines.tags.tagList=("${value}"))`;
+        return `${FILTERS.TAGS}==${value} or invoiceLines.tags.tagList==${value}`;
       },
       [FILTERS.FUND_CODE]: buildArrayFieldQuery.bind(null, ['invoiceLines.fundDistributions']),
       [FILTERS.EXPENSE_CLASS]: buildArrayFieldQuery.bind(null, ['invoiceLines.fundDistributions']),